+++ CHANGELOG.rst
@@ -3,6 +3,13 @@ Changelog
 =========
 
 
+1.8.3 (2026-06-18)
+==================
+
+- Broker graceful shutdown with task tracking
+- Broker fixes and updates
+- Signal and queue hash over name
+
 1.8.2 (2026-06-11)
 ==================